# codeclimate: less memory allocation