Use json encoder to write codeclimate report.