@Lasiar 2023-06-04
@ldez

> Hello,
> 
> do you have a problem with the code climate reporter? Can you explain your problem?

Hello, I don't have any problem, this PR just reduce allocation and memory usage.

[Source code of benchmark.](https://gist.github.com/Lasiar/c089288acfab500defda4821f24831cc)

<details>
  <summary>result on my machine:</summary>
  
`go test -run='^$' -bench=. -count=10`

| name       | old time/op | new time/op | delta   |                 |
| ---------- | ----------- | ----------- | ------- | --------------- |
| Climate-10 | 2.79µs ± 1% | 2.56µs ± 1% | \-8.00% | (p=0.000 n=9+8) |

| name       | old alloc/op | new alloc/op | delta    |                  |
| ---------- | ------------ | ------------ | -------- | ---------------- |
| Climate-10 | 3.36kB ± 1%  | 2.23kB ± 3%  | \-33.66% | (p=0.000 n=10+9) |


| name       | old alloc/op | new alloc/op | delta    |                  |
| ---------- | ------------ | ------------ | -------- | ---------------- |
| Climate-10 | 34.0 ± 0%  | 31.0 ± 0%  | \-8.82% | (p=0.000 n=10+10)|

</details>