@ldez 2023-06-04
Hello,

The problem with this kind of benchmark is in the real interest of this optimization.

The basic rule of optimization is to do it only when you have a performance problem.
There is a name for that: premature optimization.

In a general way, `json.Marshal` or `json.Unmarshal` produce more useful errors than `json.NewDecoder` and `json.NewEncoder`.

In the current situation, there's little chance of an error, so we can accept this PR.
But I want to be clear that performance optimization should be led by performance problems and not micro-benchmarks.